\name{coef-methods}
\docType{methods}
\alias{coef-methods}
\alias{coef,linCombOrBackTrans-method}
\alias{coef,unmarkedEstimate-method}
\alias{coef,unmarkedFit-method}
\title{Methods for Function coef in Package `unmarked'}
\description{
Extract coefficients
}
\section{Methods}{
\describe{

\item{object = "linCombOrBackTrans"}{Object from linearComb}

\item{object = "unmarkedEstimate"}{unmarkedEstimate object}

\item{object = "unmarkedFit"}{Fitted model}
}}
\keyword{methods}

