% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_state_functions.R
\name{doModl}
\alias{doModl}
\title{Computes detection probability matrix}
\usage{
doModl(mnum, modname1, modname2, modname3)
}
\description{
This function runs a model specified by name, where name is of the form psi(X),phi(Y),p(Z), where:
\itemize{
	\item{X is a design-matrix file for initial occupancy (psi), and specified by by the variable, modname1}
	\item{Y is a design-matrix file for state transitions (phi), and specified by by the variable, modname2}
	\item{Z is a design-matrix file for detection probabilities (p), and specified by by the variable, modname3}
}
\itemize{
\item{R = number of real parameters (psi, phi, p) and}
\item{C = number of beta's}
}
To make input simpler, the design matrix is split into 3 parts:
\itemize{
\item{Part 1 contains all rows and the columns which affect initial psi.}
\item {Part 2 contains all rows and the columns which affect transitions (e's and g's).}
\item{Part 3 contains all rows and the columns which affect detection (p's).}
}
The 3 parts are then horizontally concatenated to produce a single design matrix.
}
\references{
Miller, D. A. W., C. S. Brehem, J. E. Hines, J. D. Nichols, and R. N. Fisher. 2012. Joint estimation of habitat dynamics and species interactions: disturbance reduces co-occurrence of non-native predators with an endangered toad. Journal of Applied Ecology 81:1288-1297.
}

