\name{doublePiFun}
\alias{doublePiFun}
\title{Compute the cell probabilities for the observation classes...}
\usage{doublePiFun(p)}
\description{Compute the cell probabilities for the observation classes
in double observer sampling.}
\details{This function and \link{removalPiFun} provide example functions for computing multinomial
cell probabilities for \link{multinomPois}.}
\value{M x 3 matrix of cell probabilties for double observer observation classes.}
\arguments{\item{p}{M x 2 matrix of detection probabilities at each occasion for each site}}
