\name{fitted-methods}
\docType{methods}
\alias{fitted-methods}
\alias{fitted,unmarkedFit-method}
\alias{fitted,unmarkedFitColExt-method}
\alias{fitted,unmarkedFitOccu-method}
\alias{fitted,unmarkedFitOccuRN-method}
\alias{fitted,unmarkedFitPCount-method}
\title{Methods for Function fitted in Package `unmarked'}
\description{Methods for function \code{fitted} in Package `unmarked'
}
\section{Methods}{
\describe{

\item{object = "unmarkedFit"}{A fitted model object}

\item{object = "unmarkedFitColExt"}{A fitted model object}

\item{object = "unmarkedFitOccu"}{A fitted model object}

\item{object = "unmarkedFitOccuRN"}{A fitted model object}

\item{object = "unmarkedFitPCount"}{A fitted model object}
}}
\value{Returns a matrix of expected values}
\keyword{methods}

