\name{getB-methods}
\docType{methods}
\alias{getB}
\alias{getB-methods}
\alias{getB,unmarkedFitOccuFP-method}
\title{Methods for Function getB in Package `unmarked'}
\description{
Methods for function \code{getP} in Package `unmarked'. These methods
return a matrix of probabilities detections were certain for occupancy models
that account for false positives.
}
\keyword{methods}