\name{getFP-methods}
\docType{methods}
\alias{getFP}
\alias{getFP-methods}
\alias{getFP,unmarkedFitOccuFP-method}
\title{Methods for Function getFP in Package `unmarked'}
\description{
Methods for function \code{getP} in Package `unmarked'. These methods
return a matrix of false positive detection probabilities.
}
\keyword{methods}