\name{getN-methods}
\docType{methods}
\alias{getN}
\alias{getN-methods}
%\alias{getN,unmarkedFit-method}
\alias{getN,unmarkedFitOccu-method}

\title{Methods for Function getN in Package `unmarked'}
\description{
Methods for function \code{getN} in Package `unmarked'. These methods
return derived abundance from models that support it.
}
\section{Methods}{
\describe{
%\item{object = "unmarkedFit"}{A fitted model object}
\item{object = "unmarkedFitOccu"}{A fitted model object}
}}
\keyword{methods}
