% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_state_functions.R
\name{initPsi}
\alias{initPsi}
\title{Computes initial state vector}
\usage{
initPsi(beta, site_type)
}
\description{
Computes initial state vector
}
\references{
Miller, D. A. W., C. S. Brehem, J. E. Hines, J. D. Nichols, and R. N. Fisher. 2012. Joint estimation of habitat dynamics and species interactions: disturbance reduces co-occurrence of non-native predators with an endangered toad. Journal of Applied Ecology 81:1288-1297.
}

