% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_state_functions.R
\name{Like}
\alias{Like}
\title{Computes detection probability matrix}
\usage{
Like(beta)
}
\description{
Here's the Likelihood function for the Multi-state, integrated habitat model. It's suprisingly simple, using matrix multiplication.  It's done as a standard Multi-strata capture-recapture model. This function computes likelihood value for given betas.
}
\references{
Miller, D. A. W., C. S. Brehem, J. E. Hines, J. D. Nichols, and R. N. Fisher. 2012. Joint estimation of habitat dynamics and species interactions: disturbance reduces co-occurrence of non-native predators with an endangered toad. Journal of Applied Ecology 81:1288-1297.
}

