\name{linearComb-methods}
\docType{methods}
\alias{linearComb}
\alias{linearComb-methods}
%%\alias{linearComb,unmarkedEstimate,matrixOrVector-method}
%%\alias{linearComb,unmarkedFit,matrixOrVector-method}
\title{Methods for Function linearComb in Package `unmarked'}
\description{
Methods for function \code{linearComb} in Package `unmarked'
}
\section{Methods}{
\describe{

\item{obj = "unmarkedEstimate", coefficients = "matrixOrVector"}{Typically 
	called internally}

\item{obj = "unmarkedFit", coefficients = "matrixOrVector"}{Returns predicted
values from a model. Covariate values must be supplied.}
}}
\keyword{methods}

