\name{linearComb-unmarkedEstimate}
\title{Compute linear combinations of estimates in unmarkedEstimate objects.}
\description{Compute linear combinations of estimates in unmarkedEstimate objects.}
\details{This function computes the linear combination of parameter estimates in
\code{obj} given by the coefficient vector.  This may be useful to estimate
quantities of interest from a fitted model or to test a hypothesis.}
\alias{linearComb,unmarkedEstimate-method}
\value{an unmarkedEstimate object}
\arguments{\item{obj}{an unmarkedEstimate object}
\item{coefficients}{vector of same length as obj}}
