\name{linearComb}
\alias{linearComb}
\title{Compute linear combinations of estimates in unMarkedEstimate objects.}
\description{Compute linear combinations of estimates in unMarkedEstimate objects.}
\details{This function computes the linear combination of parameter estimates in
\code{obj} given by the coefficient vector.  This may be useful to estimate
quantities of interest from a fitted model or to test a hypothesis.}
\value{an unMarkedEstimate object}
\arguments{\item{obj}{an unMarkedEstimate object}
\item{coefficients}{vector of same length as obj}}
