% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcount.R
\name{pcount}
\alias{pcount}
\title{Fit the N-mixture point count model}
\usage{
pcount(formula, data, K, mixture = c("P", "NB", "ZIP"), starts,
  method = "BFGS", se = TRUE, engine = c("C", "R"), ...)
}
\description{
Fit the N-mixture point count model
}

