\name{predict-methods}
\docType{methods}
\alias{predict-methods}
\alias{predict,ANY-method}
\alias{predict,unmarkedFit-method}
\alias{predict,unmarkedFitColExt-method}
\alias{predict,unmarkedFitGMM-method}
\alias{predict,unmarkedFitGDS-method}
\alias{predict,unmarkedFitPCO-method}
\alias{predict,unmarkedFitList-method}
\title{ Methods for Function predict in Package `unmarked' }
\description{
These methods return predicted values from fitted model objects.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "unmarkedFit")}}{
Type must be either `state' or `det'.
}
\item{\code{signature(object = "unmarkedFitColExt")}}{
Type must be either `state' or `det'.
}

\item{\code{signature(object = "unmarkedFitGMM")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "unmarkedFitList")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
