\name{SE-methods}
\docType{methods}
\alias{SE}
\alias{SE-methods}
\alias{SE,linCombOrBackTrans-method}
\alias{SE,unmarkedEstimate-method}
\alias{SE,unmarkedFit-method}
\title{Methods for Function SE in Package `unmarked'}
\description{Extract standard errors of parameter estimates from a fitted model.}
\section{Methods}{
\describe{
\item{obj = "linCombOrBackTrans"}{A model prediction}
\item{obj = "unmarkedEstimate"}{See \code{\link{unmarkedEstimate-class}}}
\item{obj = "unmarkedFit"}{A fitted model}
}}
\keyword{methods}
