\name{SE}
\alias{SE}
\title{Compute standard error of an unMarkedEstimate object.}
\description{Compute standard error of an unMarkedEstimate object.}
\details{This function computes the large-sample standard error from the inverse of the
hessian matrix.}
\value{vector of the standard error(s) of estimates in obj}
\arguments{\item{obj}{unMarkedEstimate whose standard error is returned}}
