\name{shinyPower}
\alias{shinyPower}
\title{Launch a Shiny app to help with power analysis}
\description{
  Launch a Shiny app to test power under various scenarios. Requires the Shiny 
  package to be installed.
}
\usage{
shinyPower(object, ...)
}
\arguments{
  \item{object}{A template \code{unmarkedFit} object; see 
   documentation for \code{powerAnalysis} for details on how to create this}
  \item{...}{Currently ignored}
}
\value{No return value, called for its side effects.}
