\name{SSE}
\alias{SSE}
\title{Compute Sum of Squared Residuals for a Model Fit.}
\description{
Compute the sum of squared residuals for an unmarked fit object.  This
is useful for a \code{\link{parboot}}.
}
\usage{
SSE(fit)
}
\arguments{
  \item{fit}{An unmarked fit object.}
}
\value{
A numeric value for the models SSE.}
\seealso{
\code{\link{parboot}}
}


