\name{unmarkedEstimateList-class}
\Rdversion{1.1}
\docType{class}
\alias{unmarkedEstimateList-class}
\alias{names,unmarkedEstimateList-method}
\alias{show,unmarkedEstimateList-method}
\alias{summary,unmarkedEstimateList-method}
\title{Class "unmarkedEstimateList"}
\description{Class to hold multiple unmarkedEstimates in an
  \code{\link{unmarkedFit}}
}
\section{Slots}{
	 \describe{
    \item{\code{estimates}:}{A \code{"list"} of models.}
  }
}
\keyword{classes}
