\name{Virtual classes}
\Rdversion{1.1}
\docType{class}
\alias{optionalCharacter-class}
\alias{optionalDataFrame-class}
\alias{optionalMapInfo-class}
\alias{optionalMatrix-class}
\alias{optionalNumeric-class}
\title{Virtual classes}
\description{Virtual classes. Not to be used by users}
\section{Objects from the Classes}{No objects may be created from them.}
\section{Methods}{
No methods defined.
}
\keyword{classes}
