\name{wt}
\alias{wt}
\alias{wt.y}
\alias{wt.data}
\alias{wt.obs}
\alias{widewt}
\docType{data}
\title{Willow tit (Poecile montanus) occurence data}
\description{
Data from 237 plots surveyed on 3 occasions.
}
\usage{data(wt)}
\format{
  A data frame with 237 observations on the following 3 variables.
  \describe{
    \item{\code{elev}}{a numeric vector}
    \item{\code{forest}}{a numeric vector}
    \item{\code{length}}{a numeric vector}
  }
}
\details{
  Put details here.
}
\source{
  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
  ~~ possibly secondary sources and usages ~~
}
\examples{
data(wt)
## maybe str(wtdata) ; plot(wtdata) ...
}
\keyword{datasets}
