\name{getP-methods}
\docType{methods}
\alias{getP}
\alias{getP-methods}
\alias{getP,unmarkedFit-method}
\alias{getP,unmarkedFitOccuFP-method}
\alias{getP,unmarkedFitOccuMulti-method}
\alias{getP,unmarkedFitOccuMS-method}
\alias{getP,unmarkedFitOccuTTD-method}
\alias{getP,unmarkedFitDS-method}
\alias{getP,unmarkedFitMPois-method}
\alias{getP,unmarkedFitColExt-method}
\alias{getP,unmarkedFitPCO-method}
\alias{getP,unmarkedFitGMM-method}
\alias{getP,unmarkedFitGDS-method}
\alias{getP,unmarkedFitGPC-method}
\alias{getP,unmarkedFitDSO-method}
\alias{getP,unmarkedFitMMO-method}
\title{Methods for Function getP in Package `unmarked'}
\description{
Methods for function \code{getP} in Package `unmarked'. These methods
return a matrix of detection probabilities.
}
\section{Methods}{
\describe{
\item{object = "unmarkedFit"}{A fitted model object}
\item{object = "unmarkedFitDS"}{A fitted model object}
\item{object = "unmarkedFitMPois"}{A fitted model object}
\item{object = "unmarkedFitGMM"}{A fitted model object}
}}
\keyword{methods}
