\name{getP-methods}
\docType{methods}
\alias{getP-methods}
\alias{getP,unmarkedFit-method}
\alias{getP,unmarkedFitDS-method}
\alias{getP,unmarkedFitMPois-method}
\title{Methods for Function getP in Package `unmarked'}
\description{
Methods for function \code{getP} in Package `unmarked'. These methods return a 
matrix of detection probabilities.
}
\section{Methods}{
\describe{

\item{object = "unmarkedFit"}{A fitted model object}

\item{object = "unmarkedFitDS"}{A fitted model object}

\item{object = "unmarkedFitMPois"}{A fitted model object}
}}
\keyword{methods}

