\name{simulate-methods}
\docType{methods}
\alias{simulate-methods}
\alias{simulate,unmarkedFitColExt-method}
\alias{simulate,unmarkedFitDS-method}
\alias{simulate,unmarkedFitMPois-method}
\alias{simulate,unmarkedFitOccu-method}
\alias{simulate,unmarkedFitOccuRN-method}
\alias{simulate,unmarkedFitPCount-method}
\title{Methods for Function simulate in Package `unmarked'}
\description{
Simulate data from a fitted model.
}
\section{Methods}{
\describe{

\item{object = "unmarkedFitColExt"}{see \code{\link{colext}}}

\item{object = "unmarkedFitDS"}{see \code{\link{distsamp}}}

\item{object = "unmarkedFitMPois"}{see \code{\link{multinomPois}}}

\item{object = "unmarkedFitOccu"}{see \code{\link{occu}}}

\item{object = "unmarkedFitOccuRN"}{see \code{\link{occuRN}}}

\item{object = "unmarkedFitPCount"}{see \code{\link{pcount}}}
}}
\keyword{methods}

