\name{vcov-methods}
\docType{methods}
\alias{vcov-methods}
\alias{vcov,linCombOrBackTrans-method}
\alias{vcov,unmarkedEstimate-method}
\alias{vcov,unmarkedFit-method}
\title{Methods for Function vcov in Package `unmarked'}
\description{Extract variance-covariance matrix from a fitted model.}
\section{Methods}{
\describe{
\item{object = "linCombOrBackTrans"}{See \code{\link{linearComb-methods}}}
\item{object = "unmarkedEstimate"}{See \code{\link{unmarkedEstimate-class}}}
\item{object = "unmarkedFit"}{A fitted model}
}}
\keyword{methods}
